refactor(domain): decompose entities.py into modular package by provider
Split the monolithic entities.py (569 LOC) into a modular package structure:

- domain/entities/__init__.py - Re-exports all entities for backward compatibility
- domain/entities/base.py - BaseEntity (34 LOC)
- domain/entities/chembl_activity.py - Activity, Assay (195 LOC)
- domain/entities/chembl_structures.py - Document, Target, TargetComponent, Molecule (234 LOC)
- domain/entities/pubchem.py - Compound (37 LOC)
- domain/entities/uniprot.py - Protein (44 LOC)
- domain/entities/pubmed.py - Publication (65 LOC)

All files are now under 300 LOC limit. Existing imports like
`from bioetl.domain.entities import Activity` continue to work.
